Implement function support
This change scaffolds function support, and implements a few key functions.

Functions should have signatures that match the AMPscript signature as
closely as possible.

Functions all live within the RuntimeContext partial class.  This was
to allow functions access to the runtime context, but without needing to
introduce complexity for which runtime context a function should access,
especially if there were multiple runtime contexts available in the domain through
optimizations such as multi-threaded subscriber rendering.

Functions that have support brought in are:
* V
* OUTPUT
* OUTPUTLINE
* CONCAT
* INDEXOF