Scaffold data extension support through SQLite
This change brings in a new assembly, named Sage.Engine.Data.  This
assembly is meant to host the interface and included data extension
client to support the data extension functions in AMPscript.

This client includes an in-memory SQLite database implementation.

There is a distinct separation between the public API in the Commands
folder and what is supported in the SQLite folder in order to support
further clients at a future time.

The query builder for SQLite is as minimal as possible without needing
to bring in too many external dependencies outside of .net 7. It
explicitly separates attacker-controlled string inputs from sanitized
inputs such as QuotedIdentifier.