Enable output parameters from AMPscript Functions
Some AMPscript functions allow "output" variables that data is written to before the function returns.
An example is CONTENTBLOCKBYNAME, where the "Success" parameter is a variable that can
output.

The original AMPscript function translation works by boxing all types into 'object', instead of some other
specific type. This worked until a need for output parameters came into place.  Without modification,
if that output 'object' type was a boxed 'bool' variable - there would be no way to know what the
name of the parameter that the content wanted updated.  It would only be potentially possible if it
were passed by ref instead of by value - but even that poses challenges for optional parameters & generating
the right 'ref variable' C# syntax.

The solution was to implement a new object type that boxes each _variable_.  In this way, when the
parameter is passed in, it can be marked as a variable. This variable as part of the box has a name,
and then the variable can be updated as part of executing this function so that when it returns, it has