Implement support for referenced content
This change brings in support for CONTENT functions and embedding content.
An IContentClient is responsible for providing a file on disk that represents
the content requested, then the dynamically compiles and executes this content.

The results of the embedded content are only output to the parent content if
it existed in an inline amp block %%==%%, or from OUTPUT statements.
To support this, stack frames have been implemented.

These stack frames will help in debugging and error information as well.