Create webpage when content exceptions are caught
Instead of relying on the built-in ASP.net page, a new page named
`Exception.ampscript` is created that allows us to style the exception
page using LWC.

To facilitate the information in this page, stack frame and exception
information must make its way into the context for rendering the page.
I decided to use a JSON object in the subscriber context to bring it over,
then the author of the exception page can use it like the data came from
a data extension.

Removed async from the content client - it's clumsy and unused. Async
exceptions are aggregate exceptions instead of the thrown exception,
which makes exception handling frustrating.

Other changes:
* Minor improvement to thrown exceptions
* The provided example is removed from the web project.