Add additional AMPscript functions
Adds the following functions:
URLENCODE - Modifies a string to only include characters that are safe to use in URLs.

BUILDROWSETFROMSTRING - Creates a rowset from a character string by splitting the string at the specified delimiter.

FORMATNUMBER - Formats a number as a numeric type, such as a decimal, date, or currency value.

There is a known issue with FORMATNUMBER for rounding digits.

Specifically:
https://learn.microsoft.com/en-us/dotnet/standard/base-types/standard-numeric-format-strings

On .NET Framework and .NET Core up to .NET Core 2.0, the runtime selects the result with the greater least significant digit (that is, using MidpointRounding.AwayFromZero).
On .NET Core 2.1 and later, the runtime selects the result with an even least significant digit (that is, using MidpointRounding.ToEven).

The in-production AMPscript implementation is on .net framework, this is .net 8.